Updated temperature test

+++ mBapp/app/src/main/java/com/example/mbapp_androidapp/common/classes/ConnectedThread.kt
@@ -14,7 +14,7 @@ import java.io.IOException
 import java.io.InputStream
 import java.io.OutputStream
 
-class ConnectedThread(socket: BluetoothSocket,private val activity: MainActivity) : Thread() {
+class ConnectedThread(socket: BluetoothSocket, private val activity: MainActivity) : Thread() {
 
     private var tag = "ConnectedThread"
     private var mmInputStream: InputStream? = null
@@ -24,7 +24,7 @@ class ConnectedThread(socket: BluetoothSocket,private val activity: MainActivity
 
     companion object{
         @Volatile private var INSTANCE: ConnectedThread? = null
-        fun getActualThread(socket: BluetoothSocket?,activity: MainActivity?): ConnectedThread {
+        fun getActualThread(socket: BluetoothSocket?, activity: MainActivity?): ConnectedThread {
             return INSTANCE ?: synchronized(this) {
                 val instance = ConnectedThread(socket!!,activity!!)
                 INSTANCE = instance
@@ -68,7 +68,7 @@ class ConnectedThread(socket: BluetoothSocket,private val activity: MainActivity
                             0 -> {
                                 val sys = System.getInstance()
                                 val data = filterMsg(msg.obj.toString())
-                                sys.updateTemperature(data[0])
+                                sys.updateTemperature(data[0].toInt())
                                 sys.doorIsOpen=data[1]!="1"
                                 sys.setWeights(data[2].toInt(),data[3].toInt())
                             }
